+++ Scripts/PowerShell/Invoke-MSGraphAPI/CHANGELOG.md
@@ -1,11 +1,10 @@
-# Get-MSCloudLicenseUsage release history
+# Invoke-MSGraphAPI release history
 
 ## 1.1.0 - 2024-07-10
 
 * Renamed `Get-MsGraphAccessToken` to `Get-MsGraphAPIAccessToken` to better reflect the function's purpose.
 * Fixed a bug with the `Invoke-MSGraphAPI` pagination handling. The script now correctly handles the `@odata.nextLink` property and fetches all pages of the response.
 * Fixed the `Error Message` in the `Invoke-MSGraphAPI` function to correctly display the error message from the response body.
-* Updated  script name in the `Get-Help` output to reflect the correct script name.
 
 ## 1.0.1 - 1.0.3 (Ferry Bodijn)
 
@@ -16,4 +15,4 @@
 
 ### First version
 
-* Checks the license usage of the specified license. If the available amount is under the specified MinimumLicenseThreshold it will send a message to a slack webhook.
+* Invokes the Microsoft Graph API with paging support.